requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flask-Migrate
flask_cors
python-dotenv